Droit

+++ Algorithmique_Programmation_et_Complexité/TP9/C++/main.cpp
@@ -22,13 +22,13 @@ int main(int argc, char** argv)
 
     string src = argv[1];
     string dest = argv[2];
-    int alpha = stoi(argv[3]);
+    float alpha = stof(argv[3]);
 
     // Sigma est de l'ordre de l'écart type de la dispersion des valeurs autour du noir et du blanc
     // noir -> [0-127], blanc -> [128-255], on peut partir sur sigma = 64 par défaut
-    int sigma = 64;
+    float sigma = 64;
     if (argc == 5)
-        sigma = stoi(argv[4]);
+        sigma = stof(argv[4]);
 
     Graphe g = Graphe(src, sigma, alpha); // Utilisation du constructeur implicite string → image
     cout << "Image de taille ";
